+++ apps/sasquatch/src/projectDependency/java/com/microsoft/appcenter/sasquatch/activities/storage/TestDocument.java
@@ -1,0 +1,7 @@
+package com.microsoft.appcenter.sasquatch.activities.storage;
+
+public class TestDocument {
+
+    @SuppressWarnings("unused")
+    String key;
+}
